add function for isValidAccount and the tests
